+++ src/maven_central/mod.rs
@@ -1,4 +1,3 @@
-mod dependancies;
 mod get;
 mod get_maven;
 mod maven_error;
@@ -6,8 +5,6 @@ mod maven_error;
 pub mod metadata;
 pub mod pom;
 
-pub use dependancies::MavenDependancy;
-pub use dependancies::get_maven_dependancies;
 pub use get::get_artifact_metadata;
 pub use get::get_jar;
 pub use get::get_pom;
